Documentation on controller actions, refactor some directories
