+++ .gitignore
@@ -1,3 +1,4 @@
 tmn-api.json
-.sh
+*.sh
 .idea
+main
