+++ controllers/root.go
@@ -12,6 +12,22 @@ type RootResponse struct {
 	Msg string
 }
 
+/*********************************************************************************************************************
+ * Controller Action: GetRoot
+ * -----------------------------
+ * HTTP Method: GET
+ * API Endpoint: /
+ *
+
+ * Description:
+ * ------------
+ * Returns a simple message stating that the API is reachable.
+ *
+ * Example:
+ * --------
+ * curl -X GET http://localhost:8080. This will show the a basic message, "I am TMN-API."
+
+ *********************************************************************************************************************/
 func GetRoot(recorder *httptest.ResponseRecorder, request *http.Request, db *gorm.DB) error {
 	rootMsg := RootResponse{Msg: "I am TMN-API"}
 
@@ -21,7 +37,6 @@ func GetRoot(recorder *httptest.ResponseRecorder, request *http.Request, db *gor
 	}
 
 	recorder.Write(json)
-
 	return nil
 }
