+++ controllers/users.go
@@ -12,6 +12,23 @@ import (
 	"net/http/httptest"
 )
 
+
+/*********************************************************************************************************************
+ * Controller Action: ShowUser
+ * -----------------------------
+ * HTTP Method: GET
+ * API Endpoint: /users/{id}
+ *
+
+ * Description:
+ * ------------
+ * Returns a single JSON object representing the user with id of {id}
+ *
+ * Example:
+ * --------
+ * curl -X GET http://localhost:8080/users/1. This will show the JSON representation of the user with ID=1
+
+ *********************************************************************************************************************/
 func ShowUser(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.DB) error {
 
 	vars := mux.Vars(r)
@@ -29,6 +46,22 @@ func ShowUser(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.DB)
 	return nil
 }
 
+/*********************************************************************************************************************
+ * Controller Action: GetUsers
+ * -----------------------------
+ * HTTP Method: GET
+ * API Endpoint: /users
+ *
+
+ * Description:
+ * ------------
+ * Returns an array of all the users in the table
+ *
+ * Example:
+ * --------
+ * curl -X GET http://localhost:8080/users
+
+ *********************************************************************************************************************/
 func GetUsers(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.DB) error {
 
 	users := make([]models.User, 0)
@@ -43,13 +76,30 @@ func GetUsers(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.DB)
 	return nil
 }
 
+/*********************************************************************************************************************
+ * Controller Action: CreateUser
+ * -----------------------------
+ * HTTP Method: POST
+ * API Endpoint: /users
+ *
+
+ * Description:
+ * ------------
+ * Creates a new user with the given JSON data in the request body
+ *
+ * Example:
+ * --------
+ * curl -X POST http://localhost:8080/users -d '{"Name": "Test"}'. This will create a new user with name "Test".
+ * Note: the ID of the user is automatically indexed and handled by the ORM.
+
+ *********************************************************************************************************************/
 func CreateUser(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.DB) error {
 
 	var user models.User
 
 	err := json.NewDecoder(r.Body).Decode(&user)
 	if err != nil {
-		return errors.New("Error decoding JSON data: " + err.Error())
+		return errors.New("Error decoding JSON request body: " + err.Error())
 	}
 	defer r.Body.Close()
 
@@ -63,6 +113,22 @@ func CreateUser(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.D
 	return nil
 }
 
+/*********************************************************************************************************************
+ * Controller Action: UpdateUser
+ * -----------------------------
+ * HTTP Method: PUT
+ * API Endpoint: /users/{id}
+ *
+
+ * Description:
+ * ------------
+ * Updates the user with given {id}, any fields included in the json request body will be overwritten on the given user.
+ *
+ * Example:
+ * --------
+ * curl -X PUT http://localhost:8080/users/2 -d '{"Name": "Test1234"}' will only update the Name field of the user.
+
+ *********************************************************************************************************************/
 func UpdateUser(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.DB) error {
 
 	vars := mux.Vars(r)
@@ -70,15 +136,15 @@ func UpdateUser(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.D
 
 	id, err := strconv.Atoi(idStr)
 	if err != nil {
-		log.Println("Error converting string to int:", err)
+		log.Println("Error converting string to integer:", err)
 	}
-	user := models.User{
+	user := models.User {
 		Id: uint(id),
 	}
 
 	err = json.NewDecoder(r.Body).Decode(&user)
 	if err != nil {
-		return errors.New("Error decoding JSON data: " + err.Error())
+		return errors.New("Error decoding JSON request body: " + err.Error())
 	}
 	defer r.Body.Close()
 
@@ -92,13 +158,29 @@ func UpdateUser(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.D
 	return nil
 }
 
+
+/*********************************************************************************************************************
+ * Controller Action: DeleteUser
+ * -----------------------------
+ * HTTP Method: DELETE
+ * API Endpoint: /users/{id}
+
+ * Description:
+ * ------------
+ * Finds the user with {id}, deletes them and returns a json response indicating any errors
+
+ * Example:
+ * --------
+ * curl -X DELETE http://localhost:8080/users/2 This will delete the user with an ID of 2.
+
+ *********************************************************************************************************************/
 func DeleteUser(recorder *httptest.ResponseRecorder, r *http.Request, db *gorm.DB) error {
 
 	vars := mux.Vars(r)
 	id := vars["id"]
 
 	var user models.User
-	db.First(&user, id) // find the user with the given ID from http URL parameter (http://localhost:8080/users/{id})
+	db.First(&user, id)
 
 	result := db.Delete(&user)
 	json, err := json.Marshal(&result)
